refactor: generalize curve test
